% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMIdat.R
\docType{data}
\name{BMIdat}
\alias{BMIdat}
\title{Example Two-Sample Summary Data for use in MR Practicals.}
\format{A data frame with 62 rows and 43 variables, created using code presented in \code{vignette("MRBase")}.}
\source{
\itemize{
\item \url{https://www.nature.com/articles/nature14177}
\item \url{http://www.nealelab.is/blog/2017/9/11/details-and-considerations-of-the-uk-biobank-gwas}
}
}
\usage{
BMIdat
}
\description{
A dataset containing summary data for 62 independent SNPs associated with body mass index from Locke (2015), and summary estimates for their corresponding
association with systolic blood pressure from the UK Biobank.
}
\details{
BMIdat.
}
\author{
Wes Spiller; Jack Bowden
}
\keyword{datasets}
